adding system exit status checks on revokation

+++ lib/pkild/Model/Certificates.pm
@@ -582,18 +582,24 @@ sub revoke_user_certificate{
     my $user_cert_dir=$self->user_cert_dir($session);
     my $pdir = $self->user_parent_cert_dir($session);
     system("/usr/bin/openssl ca -revoke $user_cert_dir/crt -keyfile $pdir/private/key -cert $pdir/pem -config $pdir/openssl.cnf");
-     print STDERR "-=[$!]=-\n" if $!;
-
-    # update the Certificate Revocation list
-    system("/usr/bin/openssl ca -gencrl -keyfile $pdir/private/key -cert $pdir/pem -config $pdir/openssl.cnf -out $pdir/crl");
-    opendir(my $dh, "$user_cert_dir");
-    my @files = readdir($dh);
-    foreach my $file (@files){
-        unlink("$user_cert_dir/$file");
+    if($? == 0){
+        # update the Certificate Revocation list
+        system("/usr/bin/openssl ca -gencrl -keyfile $pdir/private/key -cert $pdir/pem -config $pdir/openssl.cnf -out $pdir/crl");
+        if($? == 0){
+            opendir(my $dh, "$user_cert_dir");
+            my @files = readdir($dh);
+            foreach my $file (@files){
+                unlink("$user_cert_dir/$file");
+            }
+            closedir $dh;
+            if( -d "$user_cert_dir"){ rmdir $user_cert_dir; };
+            if( -d "$user_cert_dir"){ print STDERR "Unable to remove $user_cert_dir\n" };
+        }else{
+            print STDERR "Unable to update the Certificate Revokation list\n";
+        }
+    }else{
+        print STDERR "Unable to revoke certificate.\n";
     }
-    closedir $dh;
-    if( -d "$user_cert_dir"){ rmdir $user_cert_dir; };
-    if( -d "$user_cert_dir"){ print STDERR "Unable to remove $user_cert_dir\n" };
     return $self;
 }
